Update erciyuan & Townscaper

+++ Assets/Effects/5_PBR/Scripts/Effects_SphereRotate.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Effects_SphereRotate : MonoBehaviour
 {
-    public int rotateSpeed = 2;
+    public float rotateSpeed = 2;
     void Start()
     {
